MCPListToolsUpdate
