+++ dev/io.openliberty.mcp.internal/bnd.bnd
@@ -42,4 +42,7 @@ WS-TraceGroup: MCP
     ../build.sharedResources/lib/junit/old/junit.jar;version=file, \
     io.openliberty.jakarta.jsonb.3.0;version=latest,\
     io.openliberty.org.eclipse.yasson.3.0;version=latest,\
-    io.openliberty.org.eclipse.parsson.1.1;version=latest
+    io.openliberty.org.eclipse.parsson.1.1;version=latest,\
+    io.openliberty.mcp;version=latest,\
+	org.json:json;version=20080701,\
+	org.skyscreamer:jsonassert
